+++ billing-service/src/main/java/com/billing/service/billing_service/repositorys/ProductsRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.billing.service.billing_service.domain.Products;
 
-public interface ProductsRepository extends JpaRepository<Products, Integer>{
+public interface ProductsRepository extends JpaRepository<Products, Long>{
 }
